move parts of cp_solver
